refactor: 온보딩 스텝 계산 및 다음 활성화 훅 분리
